cpu-o3: Various fixes to new load/store queue implementation
* Fix flag and physical addr update in SplitRequest

We use setFlags() to accumulate the flags from the sub requests
into the mainRequest. There is an assert in setFlags()  which
requires a valid Paddr. The original code updated the Paddr for
each sub request to avoid that assert. However, it failed to
check if the current sub request had a valid Paddr. This could
trigger an assertion in getPaddr().

Instead of updating the otherwise unused Paddr for each sub
request, a dummy Paddr is set now in the mainRequest before
the memory access is initiated.

* Check if the inst has squashed when translation finishes

* Properly handle delayed translations in SplitRequest

* Fix LSQRequest and LSQSenderState delete logic

The original code did not delete LSQSenderRequest when a blocked
load instruction got squashed. As a side effect, the
corresponding DynInst object did not get deleted either. The code
is now restructured to make sure that LSQSenderState always gets
deleted in the LSQRequest destructor.

* Fix status update in SplitDataRequest::finish()

The TranslationFinished flag must be set even if we get a fault
on a request. The original code did not delete the request when a
speculative load got squashed after a translation fault because
the LSQ logic thought that the translation was still ongoing (due
to the lack of the TranslitionFinished or TranslationSquashed
flag). The undeleted Request also prevented deleting the owner
binary instruction.

* Delete packets when LSQRequest gets deleted

Delete packets in the LSQRequest destructor. The original
code did not delete packets allocated in the requests which
resulted in a memory leak.

* Further fix to the LSQRequest delete logic

The destructor of the LSQRequest deletes all the packets. This fix
ensures that no packet is accessed after the dtor is called.

Change-Id: Ia7c1b02f881309506118cf0d055650fcd4b40035
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>