cpu-o3: Fix store forwarding in LSQ
The LSQ logic scans all the younger uncommitted stores when a
load is being executed to see if there is an overlapping one.
The original code recorded the starting index for the scan when
the load was inserted into the LQ. At the execution time, an
iterator was created using that recorded index. However, the
iterator was not always correctly initialized which might caused
an attempt to forward results from a older store (and a panic if
that store got squashed later).

The root of the problem is that the tail if the store queue
can wrap around in between load insertion and execution. If that
happens, we need to know whether the can-write-back iterator
(storeWBIt) has wrapped around, too, or not. Such information
cannot be deducted from the recorded starting index alone.

To fix the issue, we initialize and record an iterator at load
insertion time. The patch also removes the getBoundaryIterator()
method of the circularQueue which was only used for the store
forwarding scan (but failed to capture all the corner cases).

Change-Id: I13c4b70e9f0851c56e87288a968bd7e16f4fa18b
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>