cpu-o3: Add support for pinned writes
This patch adds support for pinning registers for a certain number of
writes.  This is only relevant for timing CPU models (functional-only
models are unaffected), and it is primarily needed to provide a
realistic execution model for micro-coded operations whose microops
can write to non-overlapping portions of a destination register, e.g.
vector gather loads.  In those cases, this mechanism can disable
renaming for a sequence of consecutive writes, thus making the
resulting execution more efficient: allocating a new physical register
for each microop would introduce a read-modify-write chain of
dependencies, while with these modifications the microops can write
back in parallel.

Change-Id: I07eb5fdbd1fa0b748c9bdc1174d9f330fda34f81
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>