arch-arm: Fix issue with SVE destructive gather loads
Fix an issue with gather loads whose destination register corresponds
to the source vector register used for address calculation.  The fix
consists in adding an extra first microop to all gather loads to copy
the source register to an auxiliary register, which is then read by
all subsequent microops instead of the original source register.

In addition, the assembly string of gather/scatter microops has been
tweaked to better describe the intent of the different operations.

Change-Id: I58a6171bfe643eedc848d0cc60d563844f9f2125
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>