cpu-o3: Restructure LSQRequest delete logic
There was a bug in the load-store forwarding logic which
might caused deleting an LSQRequest while there were
active memory request packets in flight. The following
events could trigger the problem:
1. The translation for a store in the LSQ is delayed.
2. A load with partially overlapping address range of this
   store is executed. It does not find out the address range
   conflict as the store is not completed yet.
3. The load needs to access several cache lines (SplitRequest).
   It sends out the first memory request successfully but the
   the cache gets blocked and does not accept the second
   request. The execution of load is rescheduled.
4. The store completes.
5. The cache gets unblocked and the load is executed again.
   This time however, it finds out the conflict and tries to
   delete the LSQRequest of the load - this cannot happen since
   the LSQRequest must be present when the cache response to the
   first memory request arrives back.

To avoid this problem the LSQRequest delete logic is
restructured. It also handles now the case when a partially
executed load finds out data forwarding opportunity (e.g.
the load address range is fully within that of an older
store).

Change-Id: I4b101d6a5883ff334306ded4ffa98d59a0462ddd
Signed-off-by: Gabor Dozsa <gabor.dozsa@arm.com>
Reviewed-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>