cpu-o3: Fix fault handling for blocked loads
There was a problem when a SplitRequest of a regular load
in the state of partial translation fault (i.e. some
requests complete translation successfully) started the
memory access but the cache got blocked. The executeLoad()
method erroneously set the load as 'executed' which
triggered an assertion when the cache got unblocked and
the rescheduled load got executed again.

Change-Id: I1537b6381ae984d514475fd9bdef499fe522ef7f
Signed-off-by: Gabor Dozsa <gabor.dozsa@arm.com>
Reviewed-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>