mem: Revert using read mask in the Packet class
The read mask was introduced by the "arch-arm: Add
support for first/non-faulting contiguous load" commit.
Applying it in the setData() method of the Packet class
is erroneous and makes various non-trivial SVE application
to fail in O3/FS mode.

This fix also makes sure that the write mask is applied
only if the mask and the data is of the same size.

Change-Id: Ic4a3323cc68a195dccb0d1e528f8209d43dbe8d8
Signed-off-by: Gabor Dozsa <gabor.dozsa@arm.com>