cpu: Fixed predicated false mem. insts. on timing and minor
TimingCPU fix:
The predicate and memAccPredicate flags on SimpleThread are read before
being set when using the timing simple cpu. So we make sure they are
initialized the SimpleThread constructor. Makes sure the predicate
flags in the execution context are reset in the constructor and before
executing an instruction in the timing simple cpu. This fixes two bugs:
1) random behavior due to these flags being used uninitialized; 2) when
all subsequent load/stores are not execute after the first time a predicate
flag is set to false.

MinorCPU fix:
Changed the location where completeAcc is called when a read request is not
issued. Calling it too early may overwrite data on the destination register
before it's consumed by an inflight instruction.
Also modified Minor::ExecContext to use the predicate flags stored in the
MinorDynInst instance, so these flags properly propagate when processing
instructions in multiple steps.

Change-Id: Idb043f32ba0173d264dccfc66cac48a9564e6134
Reviewed-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>
Signed-off-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>