cpu-o3: Fix squash handling for insts with pinned dest regs
An inst with pinned dest regs can get squashed and then re-renamed
in case of a memory order violation. We need to reset the pinned
counters of the physical register at the squash in order to ensure
that the re-rename will find them in a consistent state and will
pin the same physical register (if the squash only affects
some of the instructions which writes into the same pinned reg).

Change-Id: I77158fed16de6f9cdd1fd1f7ba02148b04d1c9b6
Signed-off-by: Gabor Dozsa <gabor.dozsa@arm.com>
Reviewed-by: Giacomo Gabrielli <giacomo.gabrielli@arm.com>