power: add parsing of Vector2d stats for power eq
Adding parsing of Vector2d stats (such as issue queue instruction class
counts in O3) to the equation parser.  This is a little tedious as

* Vector2d (or rather the associated Info class) does not allow access
  to the actual data; instead, one has to pretend printing the stat and
  the data is copied over into an accessible array (cvec)

* the parsing of the 2d stat names, because well.. parsing :)

Added a cache for Vector2d name to fully parsed out stat + index so that
we don't have to run through the parsing code every time we print stats.

Change-Id: I8eb821c05ee60b53a869bbd73164e91dba53400b
Signed-off-by: Stephan Diestelhorst <stephan.diestelhorst@arm.com>